requirements.txt
black==24.4.2
boto3==1.34.133
botocore==1.34.133
dynaconf==3.2.5
easydict==1.13
fastapi==0.111.0
fsspec==2023.4.0
fuzzywuzzy==0.18.0
gensim==4.2.0
ipython==8.12.3
isort==5.13.2
keybert==0.8.5
langchain==0.2.7
mlflow==2.11.3
mysql_connector_repackaged==0.3.1
nltk==3.7
numpy==1.22.0
openai==1.35.13
pandas==1.5.3
pdfkit==1.0.0
plotly==5.9.0
psycopg2_binary==2.9.6
pydantic==1.10.13
PyJWT==2.8.0
PyJWT==2.8.0
pymilvus==2.4.4
pyspark==3.5.1
python_dateutil==2.9.0.post0
python_dateutil==2.9.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.3
sentence_transformers==2.2.2
setuptools==69.5.1
spacy==3.5.3
sql_metadata==2.12.0
SQLAlchemy==2.0.31
tiktoken==0.3.3
torch==1.13.1
transformers==4.22.0
wrapt==1.16.0
wrapt_timeout_decorator==1.4.0
